Clarify Testing not part of Installation
Previous wording in Installation section suggested only a test contribution required, but the end of the Testing section seemed to suggest doing the full set of tests before putting into production. This PR clarifies the Testing section is for developers or those wanting to do more extensive tests on their instance.